Add legs, puzzle, and support for Kailh Box switches
Legs: If you want to use less material when printing you can now opt to
use a thinner case but add legs to lift it up from the ground to have
the switches not touch the floor.

Puzzle: Ever wanted to  make a MASSIVE switch holder but don't have a
MASSIVE printer? Print it in pieces and lay them in a puzzle! You can
glue them toghether if you want them to stick like that, or just hook
them in temporary.

Kailh: The Kailh box switches do not have the same side cuts as Cherry
or Gateron. This meant that the default values made it difficult to put
the Kailh switches in the holder. Updated default values now allow for
Kailh, albite still a snugger fit than Cherry or Gateron but I think it
will be problematic to support all equaly well.